api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api"
version = "1.0.0"
classifiers = ["Private :: Do Not Upload"]
dependencies = [
    "celery[redis] ~= 5.2.7",
    "django ~= 4.1.6",
    "django-environ ~= 0.9.0",
    "django-extensions ~= 3.2.1",
    "djangorestframework ~= 3.14.0",
    "psycopg2 ~= 2.9.5",
]

[project.optional-dependencies]
dev = [
    "bandit ~= 1.7.4",
    "black ~= 23.1.0",
    "celery-types ~= 0.14.0",
    "django-types ~= 0.16.0",
    "djangorestframework-types ~= 0.8.0",
    "factory-boy ~= 3.2.1",
    "ipdb ~= 0.13.9",
    "isort ~= 5.12.0",
    "mypy ~= 1.0.0",
    "pip-tools ~= 6.12.2",
    "pytest ~= 7.2.0",
    "pytest-clarity ~= 1.0.1",
    "pytest-cov ~= 4.0.0",
    "pytest-django ~= 4.5.2",
    "pytest-mock ~= 3.10.0",
    "ruff ~= 0.0.246",
    "watchdog[watchmedo] ~= 2.2.1",
    "wheel ~= 0.38.4",
]

[tool.mypy]
exclude = ["main/migrations"]
files = ["main"]
mypy_path = "stubs"
show_error_codes = true
strict = true
warn_unreachable = true
warn_unused_ignores = true

[tool.ruff]
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    "venv",
]
ignore = [
    "E501",
    "S101",
]
line-length = 88
select = [
    "A", # flake8-builtins
    "B", # flake8-bugbear
    "BLE", # flake8-blind-except
    "C4", # flake8-comprehensions
    "DJ", # flake8-django
    "DTZ", # flake8-datetimez
    "E", # pycodestyle
    "EXE", # flake8-executable
    "F", # Pyflakes
    "G", # flake8-logging-format
    "INP", # flake8-no-pep420
    "ISC", # flake8-implicit-str-concat
    "N", # pep8-naming
    "PGH", # pygrep-hooks
    "PIE", # flake8-pie
    "PL", # Pylint
    "PT", # flake8-pytest-style
    "PTH", # flake8-use-pathlib
    "PYI", # flake8-pyi
    "Q", # flake8-quotes
    "RET", # flake8-return
    "RSE", # flake8-raise
    "RUF", # Ruff-specific rules
    "S", # flake8-bandit
    "SIM", # flake8-simplify
    "T10", # flake8-debugger
    "UP", # pyupgrade
    "W", # pycodestyle
]
target-version = "py311"
